+++ Mindstorms.Core/SetMotorSpeedParams.cs
@@ -5,13 +5,13 @@ namespace Mindstorms.Core
 {
     public class SetMotorSpeedParams
     {
-        private const short MaxSpeed = 100;
-        private const short MinSpeed = -100;
+        private const sbyte MaxSpeed = 100;
+        private const sbyte MinSpeed = -100;
 
         public OutputPort OutputPort;
-        public short Speed;
+        public sbyte Speed;
 
-        public SetMotorSpeedParams(OutputPort outputPort, short speed)
+        public SetMotorSpeedParams(OutputPort outputPort, sbyte speed)
         {
             OutputPort = outputPort;
             Speed = Math.Max(Math.Min(speed, MaxSpeed), MinSpeed);
